feat: rename 'backends' to 'clients'
