+++ src/reddish/__init__.py
@@ -2,5 +2,5 @@
 from reddish._core.command import Args as Args, Command as Command
 from reddish._core.multiexec import MultiExec as MultiExec
 import reddish._core.errors as errors
-import reddish.backends as backends
+import reddish.clients as clients
 import reddish.models as models
