+++ src/reddish/clients/anyio/_client.pyi
@@ -1,5 +1,5 @@
 import anyio
-from reddish.backends._client_stubs import AsyncRedis
+from reddish.clients._client_stubs import AsyncRedis
 
 class Redis(AsyncRedis):
     def __init__(self, stream: anyio.abc.ByteStream) -> None: ...  # type: ignore
