+++ src/reddish/clients/asyncio/_client.pyi
@@ -1,5 +1,5 @@
 import asyncio
-from reddish.backends._client_stubs import AsyncRedis
+from reddish.clients._client_stubs import AsyncRedis
 
 class Redis(AsyncRedis):
     def __init__(
