+++ src/reddish/clients/socket/_client.py
