+++ src/reddish/clients/socket/_client.pyi
@@ -1,5 +1,5 @@
 import socket
-from reddish.backends._client_stubs import SyncRedis
+from reddish.clients._client_stubs import SyncRedis
 
 class Redis(SyncRedis):
     def __init__(self, stream: socket.socket) -> None: ...
