+++ src/reddish/clients/trio/_client.pyi
@@ -1,5 +1,5 @@
 import trio
-from reddish.backends._client_stubs import AsyncRedis
+from reddish.clients._client_stubs import AsyncRedis
 
 class Redis(AsyncRedis):
     def __init__(self, stream: trio.abc.Stream) -> None: ...
